UPDATE ::: Pushing updates to allow ToDatabase to work for big query
git-svn-id: svn://107.170.10.188:8080/Semoss/dev@43045 ef391c1c-ee15-45f7-b1d6-dde3d27a2169

Former-commit-id: 46511fa1bb8901e09fde6926d82a10ff2e3d8455

+++ src/prerna/ds/rdbms/RdbmsFrameBuilder.java
@@ -172,7 +172,7 @@ public class RdbmsFrameBuilder {
 							if (value != null) {
 								ps.setInt(colIndex + 1, value);
 							} else {
-								ps.setNull(colIndex + 1, java.sql.Types.DOUBLE);
+								ps.setNull(colIndex + 1, java.sql.Types.INTEGER);
 							}
 						}
 					} else if(type == SemossDataType.DOUBLE) {
